feat: Move func default from base to Unity version
The EnumFuncMap should be a one time setup to help reduce exception points due to potential bad map states.
